Move time facility to OS layer
Migrates the Time facility to the OS layer.
