+++ src/nstime.c
@@ -2,6 +2,7 @@
 
 #include "jemalloc/internal/assert.h"
 #include "jemalloc/internal/nstime.h"
+#include "jemalloc/internal/os.h"
 
 #define BILLION UINT64_C(1000000000)
 #define MILLION UINT64_C(1000000)
@@ -192,68 +193,9 @@ nstime_ms_since(const nstime_t *past) {
 	return nstime_ns_since(past) / MILLION;
 }
 
-#ifdef _WIN32
-#	define NSTIME_MONOTONIC false
-static void
-nstime_get(nstime_t *time) {
-	FILETIME ft;
-	uint64_t ticks_100ns;
-
-	GetSystemTimeAsFileTime(&ft);
-	ticks_100ns = (((uint64_t)ft.dwHighDateTime) << 32) | ft.dwLowDateTime;
-
-	nstime_init(time, ticks_100ns * 100);
-}
-#elif defined(JEMALLOC_HAVE_CLOCK_MONOTONIC_COARSE)
-#	define NSTIME_MONOTONIC true
-static void
-nstime_get(nstime_t *time) {
-	struct timespec ts;
-
-	clock_gettime(CLOCK_MONOTONIC_COARSE, &ts);
-	nstime_init2(time, ts.tv_sec, ts.tv_nsec);
-}
-#elif defined(JEMALLOC_HAVE_CLOCK_MONOTONIC)
-#	define NSTIME_MONOTONIC true
-static void
-nstime_get(nstime_t *time) {
-	struct timespec ts;
-
-	clock_gettime(CLOCK_MONOTONIC, &ts);
-	nstime_init2(time, ts.tv_sec, ts.tv_nsec);
-}
-#elif defined(JEMALLOC_HAVE_CLOCK_GETTIME_NSEC_NP)
-#	define NSTIME_MONOTONIC true
-static void
-nstime_get(nstime_t *time) {
-	nstime_init(time, clock_gettime_nsec_np(CLOCK_UPTIME_RAW));
-}
-#elif defined(JEMALLOC_HAVE_MACH_ABSOLUTE_TIME)
-#	define NSTIME_MONOTONIC true
-static void
-nstime_get(nstime_t *time) {
-	static mach_timebase_info_data_t sTimebaseInfo;
-	if (sTimebaseInfo.denom == 0) {
-		(void)mach_timebase_info(&sTimebaseInfo);
-	}
-	nstime_init(time,
-	    mach_absolute_time() * sTimebaseInfo.numer / sTimebaseInfo.denom);
-}
-#else
-#	define NSTIME_MONOTONIC false
-static void
-nstime_get(nstime_t *time) {
-	struct timeval tv;
-
-	gettimeofday(&tv, NULL);
-	nstime_init2(time, tv.tv_sec, tv.tv_usec * 1000);
-}
-#endif
-
 static bool
 nstime_monotonic_impl(void) {
-	return NSTIME_MONOTONIC;
-#undef NSTIME_MONOTONIC
+	return OS_TIME_MONOTONIC;
 }
 nstime_monotonic_t *JET_MUTABLE nstime_monotonic = nstime_monotonic_impl;
 
@@ -265,27 +207,15 @@ const char *const prof_time_res_mode_names[] = {
 };
 
 static void
-nstime_get_realtime(nstime_t *time) {
-#if defined(JEMALLOC_HAVE_CLOCK_REALTIME) && !defined(_WIN32)
-	struct timespec ts;
-
-	clock_gettime(CLOCK_REALTIME, &ts);
-	nstime_init2(time, ts.tv_sec, ts.tv_nsec);
-#else
-	unreachable();
-#endif
-}
-
-static void
 nstime_prof_update_impl(nstime_t *time) {
 	nstime_t old_time;
 
 	nstime_copy(&old_time, time);
 
 	if (opt_prof_time_res == prof_time_res_high) {
-		nstime_get_realtime(time);
+		os_time_get_realtime(time);
 	} else {
-		nstime_get(time);
+		os_time_get(time);
 	}
 }
 nstime_prof_update_t *JET_MUTABLE nstime_prof_update = nstime_prof_update_impl;
@@ -295,7 +225,7 @@ nstime_update_impl(nstime_t *time) {
 	nstime_t old_time;
 
 	nstime_copy(&old_time, time);
-	nstime_get(time);
+	os_time_get(time);
 
 	/* Handle non-monotonic clocks. */
 	if (unlikely(nstime_compare(&old_time, time) > 0)) {
